[TEST] Restructured tests
