+++ tests/_regtest_outputs/test_pyprophet_score.test_osw_0.out
