+++ tests/_regtest_outputs/test_pyprophet_score.test_osw_2.out
